* add J2 plasticity model evaluator
* add ability to update state variables such that
  state_old = state_new
* fix a few bugs in the continuation stepping code

+++ src/primal_problem.hpp
@@ -48,7 +48,7 @@ class PrimalProblem
     double gamma;
 
     double tolerance;
-    double max_iters;
+    unsigned max_iters;
 
     void compute_residual();
     void compute_jacobian();
